Added in sponsor ads, minor changes to headings, TOC
As per email, I haven't managed to get a satisfactory .epub format out
yet. So this is a zip of the pdf and the doc.

The current version hasn’t changed very much. Changes are:
- Added in the sponsor advertisements, copyright page
- Made some minor changes to the headings
- Made some changes suggested by the spell check on my PC
- Standardized on PowerShell rather than Powershell
- I’ve taken the line of dots out of the TOC. I think this looks better
because the TOC is both a table of contents and it serves as a list of
the practices